[#7][#8] refactor: reorganize components structure
