+++ src/components/admin/matching/matchingResult/MatchedMiniTable.tsx
@@ -1,7 +1,7 @@
 import { createColumnHelper, ColumnDef } from "@tanstack/react-table";
 import { MatchingResult } from "../../../../types/matching";
 import TimeTableButton from "../TimeTableButton";
-import BaseMiniTable from "../../../table/BaseMiniTable";
+import BaseMiniTable from "../tables/BaseMiniTable";
 
 interface MatchedMiniTableProps {
   data: MatchingResult[];
