+++ src/components/admin/matching/matchingResult/MatchingResultTable.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { BaseTable } from "../tables/BaseTable";
 import { MatchingResult } from "../../../../types/matching";
 import TimeTableButton from "../TimeTableButton";
-import MatchedMiniTable from "../miniTables/MatchedMiniTable";
+import MatchedMiniTable from "./MatchedMiniTable";
 import { ColumnDef } from "@tanstack/react-table";
 
 interface Props {
