+++ src/pages/admin/matching/tabs/DisabledStudentTab.tsx
@@ -5,26 +5,26 @@ import { DisabledStudent, Supporter } from "../../../../types/user";
 
 // 더미 데이터
 const dummyDisabledStudents: DisabledStudent[] = [
-  {
-    id: 1,
-    name: "장애학생1",
-    gender: "남성",
-    department: "컴퓨터공학과",
-    grade: "2학년 1학기",
-    disabilityType: "시각장애",
-    matchingStatus: "waiting",
-    matchingCandidates: []
-  },
-  {
-    id: 2,
-    name: "장애학생2",
-    gender: "여성",
-    department: "전자공학과",
-    grade: "3학년 2학기",
-    disabilityType: "청각장애",
-    matchingStatus: "waiting",
-    matchingCandidates: []
-  }
+  // {
+  //   id: 1,
+  //   name: "장애학생1",
+  //   gender: "남성",
+  //   department: "컴퓨터공학과",
+  //   grade: "2학년 1학기",
+  //   disabilityType: "시각장애",
+  //   matchingStatus: "waiting",
+  //   matchingCandidates: []
+  // },
+  // {
+  //   id: 2,
+  //   name: "장애학생2",
+  //   gender: "여성",
+  //   department: "전자공학과",
+  //   grade: "3학년 2학기",
+  //   disabilityType: "청각장애",
+  //   matchingStatus: "waiting",
+  //   matchingCandidates: []
+  // }
 ];
 
 const dummySupporters: Supporter[] = [
